# server/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Herbal CRM API
  version: 1.0.0
  description: |
    API documentation for Herbal CRM. Use the Authorize button to enter your JWT Bearer token for protected endpoints.
servers:
  - url: http://localhost:3001
  - url: https://herbal-crm-api.onrender.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        googleId: { type: string, nullable: true }
        name: { type: string }
        email: { type: string }
        password: { type: string, nullable: true }
        role: { type: string, enum: [user, admin] }
        createdAt: { type: string, format: date-time }
        isVerified: { type: boolean }
        verificationToken: { type: string, nullable: true }
        resetPasswordToken: { type: string, nullable: true }
        resetPasswordExpires: { type: string, format: date-time, nullable: true }
...
    Segment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        rules: { type: object }
        mongoQuery: { type: object, nullable: true }
        customerCount: { type: integer }
        isActive: { type: boolean }
        lastRun: { type: string, format: date-time, nullable: true }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Campaign:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        message: { type: string }
        segmentId: { type: string }
        sentAt: { type: string, format: date-time }
        customerCount: { type: integer }
        status: { type: string, enum: [draft, sent, failed] }
        error: { type: string, nullable: true }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: JWT token and user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    $ref: '#/components/schemas/User'
  /api/auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '201':
          description: Registration success
  /api/auth/verify-email:
    get:
      summary: Verify email
      parameters:
        - in: query
          name: token
          schema: { type: string }
      responses:
        '200':
          description: Email verified
  /api/auth/resend-verification:
    post:
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200':
          description: Verification email resent
  /api/auth/request-reset:
    post:
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200':
          description: Password reset email sent
...
  /api/segments:
    get:
      summary: List all segments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Segment'
    post:
      summary: Create a new segment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                rules: { type: object }
      responses:
        '201':
          description: Created
...
  /api/campaigns:
    get:
      summary: List all campaigns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
    post:
      summary: Create and send a campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                message: { type: string }
                segmentId: { type: string }
      responses:
        '201':
          description: Created
  /api/analytics/campaigns:
    get:
      summary: Get campaign analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema: { type: string, format: date }
        - in: query
          name: endDate
          schema: { type: string, format: date }
      responses:
        '200':
          description: Campaign analytics
  /api/analytics/segments:
    get:
      summary: Get segment analytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Segment analytics
  /api/analytics/campaigns/csv:
    get:
      summary: Export campaign analytics as CSV
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV file
...